Add -d and -D options to bin/sequel for dumping schema migrations
The -d option uses type translation, the -D option does not.  The
schema migration is dumped to the standard output.  This should be
helpful if you want an immediate look at the columns in the
database.  If you want to use it as the basis for an initial
migration, just redirect $stdout to a file.

While here, add a -h option to print the usage, since that is more
standard than -?.

Check for options that can't be used together, such as -D or -d
with -m.  Also check for options that require other options, such
as -M without -m.

Don't print the options when an error occurs, just print an error
message.

I initially thought of having -d/-D use -m to create a file in that
directory, but there are a lot of corner cases there that I don't
want to deal with.