Major refactoring of transaction support
This commit is a major refactoring of the transaction support in
Sequel.  Before, there was a generic transaction method in
Sequel::Database, but almost all adapters overrode that completely,
leading to a lot of repetitive adapter code.

This commit makes the transaction code generic and modular, so the
adapter only have to replace the parts they need.

This commit moves most of the transaction code into _transaction,
which all adapters share except for SQLite and Amalgalite. Those
adapters use connection transaction methods that take blocks,
which isn't compatible with the standard approach.

With the new generic transaction support, savepoint support is
easy to add.  It was extracted out of the PostgreSQL adapter into
a module that other adapters can use.  Since MySQL supports
savepoints, the shared MySQL adapter now includes savepoint support.
It was a little bit of work, but savepoint support works on MySQL
and PostgreSQL using either the native adapters, DataObjects, or
JDBC.

If other databases support savepoints, please try to include the
savepoint module in the adapter and see if it works, and send in
a patch if it does.

This needs testing on all adapters.  I've tested all adapters I have
access to (MySQL, PostgreSQL, SQLite, Amalgalite, JDBC, DataObjects),
but all other adapters need testing, especially Oracle, Firebird, and
ODBC.

Another improvement in the new generic transaction method is that
if there is an error raised when a transaction is committed, Sequel
handles it correctly now.  Before I think it was only handled
correctly in the PostgreSQL adapter.

I moved the test for savepoint support from the PostgreSQL adapter
specs to the integration tests, conditional on whether the database
being used supports savepoints.  This allows testing both on
PostgreSQL and MySQL, as well as other databases in the future.

While here I've made the transaction integration tests work on MySQL,
which requires you use the InnoDB table engine.