Add identity_map plugin, for creating temporary thread-local identity maps with some caching
The identity map plugin allows for a creating a 1-1 correspondence of
of model objects to the database rows they represent.

What an identity map does is make the following statement true:

  Album.filter{(id > 0) & (id < 2)}.first.object_id == \
    Album.first(:id=>1).object_id

In general Sequel use, that statement would be false.  However, if you
enable the identity_map plugin and wrap the code in a
Sequel::Model.with_identity_map block, it will be true.

Identity maps by themselves don't usually offer a significant
advantage, but they allow you avoid issuing database queries if
a request comes in for an object that is already in the identity map.
The identity_map plugin does that by making Model.[] use the current
identity map as a lookup cache, similar to how the caching plugin
works.

This commit also splits the private instance method
load_associated_objects into 2 methods, one that handles caching,
callbacks, and reciprocals, and another that retreives the
associated objects from the database. The new method is overridden in
the identity map_plugin to use the current identity map as a lookup
cache for many_to_one associated records.

The identity map plugin should be thought of purely as a possible
performance enhancer.  There is no strategic advantage to using it.
There are some cases where the behavior will be different depending
on whether an identity map is used (like the code above), but they
are probably uncommon.

In most cases the duration of identity maps should be very short.
The only time a long-lived identity map makes sense is if you have
a single threaded application and nothing else is modifying the
database.  Sequel doesn't offer a way to have an open ended identity
map, the only access is through the with_identity_map method that
takes a block and ensures the identity_map is destroyed when the
block exits.

Here's a Rack middleware that you can use to use a separate identity
map per request:

  Sequel::Model.plugin :identity_map
  class SequelIdentityMap
    def initialize(app)
      @app = app
    end
    def call(env)
      Sequel::Model.with_identity_map{@app.call(env)}
    end
  end