Add tactical_eager_loading plugin, similar to DataMapper's strategic eager loading
Tactical Eager Loading is basically the same thing as DataMapper's
Strategic Eager Loading.  It allows you to avoid using eager on the
dataset and still eagerly load the records.  It depends on the
identity_map plugin and only works if there is an active identity
map.

Basically, this allows the following code to issue only two queries:

  Sequel::Model.with_identity_map do
    Album.filter{id<100}.all do |a|
      a.artists
    end
  end

The plugin depends on Dataset#all being used to retreive records, so
the following will not eagerly load the artists:

  Sequel::Model.with_identity_map do
    Album.filter{id<100}.each do |a|
      a.artists
    end
  end

That is because when artists is called on the first album returned,
no other albums have been created.