Add lazy_attributes plugin, for creating attributes that can be lazily loaded from the database
The lazy_attributes plugin builds on the tactical_eager_loading and
identity_map plugins.  It allows you set an attribute as being lazily
loaded:

  Album.plugin :lazy_attributes
  Album.lazy_attributes :review

This does 2 things:

1) It removes the :review attribute from being selected by default,
   so if you do Model.first, it won't include the :review column.

2) It'll define a review instance method that will query the
   database for value of :review when called, if that value doesn't
   already exist.

Of course, if it just did the above, it wouldn't require the
identity_map or tactical_eager_loading plugins.  It requires those
plugins because if you select a group of objects that have a lazy
loaded attribute and ask for the lazy loaded attribute for one of
them, it'll get the lazy loaded attribute for all of the objects
in one query, instead of using a query for each object, similar to
how the tactical_eager_loading plugin loads associations:

  # 2 queries total
  Sequel::Model.with_identity_map do
    Album.filter{id<100}.all do |a|
      a.review
    end
  end