Keep track of loaded plugins, allows plugins to depend on other plugins
Before, if you tried to load the same plugin multiple times, Sequel
would let you, and the effects could be bizarre.  This commit makes
Sequel keep track of the plugins loaded for the class, and if an
attempt is made to load the same plugin again (or in a subclass),
Sequel will just return.

This change allows plugins to have dependences by calling
Model.plugin in the plugin's apply method.  For example:

    module LazyAttributes
      def self.apply(model)
        model.plugin :tactical_eager_loading
      end

This way, if you try to load the lazy_attributes plugin, it will load
the tactical_eager_loading plugin (which it depends on) first. This
commit includes this change to the lazy_attributes plugin, as well as
a similar one to the tatical_eager_loading plugin to depend on the
identity_map plugin.

This commit contains so other changes:

* The plugin_opts methods defined now include multiple args in an
  array if multiple args are given.  This breaks backwards
  compatibility slightly, but I think it's the correct decision.

* If a block is passed to the Model.plugin method, it is given to
  the plugin's apply method.

* You can see which plugins are loaded for the model by calling
  Model.plugins.

The plugins spec was sorely lacking.  While it did cover all points,
it was pretty monolithic and didn't test all that it should.  I
added a bunch of specific tests that should cover all behavior.

There was a change to the caching plugin specs where the same class
attempting to load the caching plugin again with a different value
for the ttl.  This no longer works because Model.plugin returns
before loading the plugin a second time.  If any of your code does
something similar, you'll have to change it.