Allow plugins to have a configure method that is called on every attempt to load them
This change should bring back backwards compatibility for many
Sequel plugins that take options.  Since apply is now only called
once, the first time the plugin is loaded, the wasn't a way to have
call the plugin method a second time with different options and
have the options of the second call be applied.

The plugin code will now call a plugin's configure method if the
plugin responds to it, every time the plugin method is called, not
just the first time.  This is useful if the call to the plugin itself
does something, as opposed to just loading code that can be used
later.  The plugins that ship with Sequel have all been modified to
support this, in the following ways:

* Caching: Can be called a second time to set a different cache or
  ttl.

* Single Table Inheritance: Can be called a second time to set a
  different column as the inheritance key.

* Lazy Attributes: Can specify the columns that are lazy directly
  on the plugin call, similar to the serialization plugin:

    Album.plugin :lazy_attributes, :review

* Serialization: This got an extensive makeover and now allows you
  to use different serialization formats for different columns.  It
  can be used like the lazy_attributes plugin, and no longer requires
  you to specify the serialized attributes on the plugin call:

    Album.plugin :serialization
    Album.serialize_attributes :marshal, :review
    Album.serialize_attributes :yaml, :name
    Album.serialization_map #{:name=>:yaml, :review=>:marshal}

  Backwards compatibility should be kept for the public API, the
  private instance methods have a new API though.