Fix MySQL command out of sync errors, disconnect from database if they occur
I've found another cause of the command out of sync errors.
Previously, If you did something like the following:

  DB << "SHOW DATABASES"

The next query would give you an error.  That's because the previous
query returned results and they weren't picked up.  To prevent that
from happen, I turned off the query_with_result setting, so it's back
to the default, which is true.

This actually simplified the logic a bit and makes it so that you
can use DB#<< and similar methods with queries that would usually
need to return results.

AFAICT, this error has been present since at least 2007.

If a commands out of sync error does occur, this disconnects the
connection using DatabaseDisconnectError.  If there are other errors
that indicate that the connection should be disconnected, please
let me know.