Make graphing a complex dataset work correctly
Before, Dataset#graph didn't work correctly if the argument was a
complex dataset (anything other than SELECT * FROM table).  This
is because it only passed the first source of the dataset to
join_table.  You could work around the issue by using from_self,
but that shouldn't be necessary.

To keep the SQL simple in the common case, check if the passed
dataset is a simple dataset and use just the table symbol in that
case.  Otherwise, pass the dataset itself.