Add :ignore_index_errors option to Database#create_table and :ignore_errors option to Database#add_index
The schema_dumper extension is going to be using these options to
make it more likely that the migrations it generates will be
loadable.  Since indexes generally have no effect on behavior
(except for unique indexes), the inability to load one shouldn't
be a reason to cancel the migration.

I'm not happy with this way of doing things, but I can't really
think of a better idea that will allow dumping of PostgreSQL tables
with indexes on text columns and attempting to restore them on
MySQL.