Don't dump partial indexes in the MySQL adapter
Previously, if you had a partial index on a text column in a MySQL
database and you tried to dump it and restore it on a MySQL database,
an error would be raised, since Sequel doesn't offer a way to create
partial indexes.

This treats partial indexes like function indexes on PostgreSQL,
by not dumping them.  If you are using partial indexes, you are
definitely in database specific territory and should not expect
perfect behavior from Sequel's migration dumper/restorer.