Make schema_dumper extension ignore errors with indexes unless it is dumping in the database-specific type format
When doing a database independent schema dump, include code to ignore
errors when creating indexes.  This should allow you to dump the
schema from a PostgreSQL database with an index on a text column and
restore it in a MySQL database.

This modifies the bin/sequel tool to use the new feature when copying
databases, and tweaks the output slightly.