Make :eager_graph association option work correctly when eagerly loading many_to_many associations
Before, this didn't work correctly because eagerly loading
many_to_many associations requires a key field in the join table be
returned, and the :eager_graph association option wiped out the
selection of that key field.

This commit has an interesting use of add_graph_aliases to map
a table column into an aliased name in the query and the graphed
output.

Fixing this issue necessitated a bit of refactoring.  Push the
necessary code into the association reflection for easy access.

While here, make the :conditions option with a placeholder literal
string work when eager loading.

Also while here, fix a bug in eager_unique_table_alias by making
it consider previously joined tables as well as tables in the FROM
clause.