Add the :cartesian_product_number option to associations, for specifying if they can cause a cartesian product
Previously this behavior was hard coded into the associations, and
the user couldn't override it.  Now it should be a number between
0 and 2:

* 0 - This association cannot cause a cartesian product, as only
  1 row exists in the join table for each row in the current table.

* 1 - This association can cause a cartesian product if used in
  the same query as another association whose cartesian product
  number is also 1, but will not cause a cartesian product if
  used by itself.

* 2 - This association causes a cartesian product even if used by
  itself.

many_to_one associations are 0 by default, and *_to_many
associations are 1 by default.  You may want to set a many_to_many
association to 2 if both of the joins it does can cause multiple
rows for each row in the current table. For example:

  Artist.one_to_many :albums
  Album.one_to_many :tracks
  Artist.many_to_many :tracks, :join_table=>:albums, :right_key=>:id,
    :right_primary_key=>:album_id, :cartesian_product_number=>2

While here, fix a description for one of the specs.