Add Dataset#qualify_to and #qualify_to_first_source, for qualifying unqualified identifiers in the dataset
Because Sequel represents SQL constructs as objects and not strings,
it is possible to go through the tree of objects that represent a
query, find all of the unqualified identifiers, and qualify them with
a table.  This is mostly useful when you have a dataset for a single
table with SELECT, WHERE, ORDER, GROUP, or HAVING clauses that use
unqualified columns that you want to join to another table/dataset.
Qualification is not needed on the main dataset, but if the table
being joined has columns with the same name as columns currently in
the dataset, you have to qualify those columns.

The main use case for this would be when you are using models that
have a filter or order by default.  If you try to join them to
another table/model where the column names overlap, before you'd
need to redo the filters or order, or make sure that the model uses
qualified columns by default.  Now, you can just call
qualify_to_first_source before you join and Sequel will take care
of things for you.

While here, allow sub_subscript to be used on all types of
expressions, not just symbols. Since functions, qualified
identifiers, and basically any other type of expression can be an
array, it makes sense to give the objects that represent those
constructs the sql_subscript method.  Also, I did some testing on
PostgreSQL and found that the subcripts themselves don't have to be
integers, they can be expressions as well, so allow any expression
to be used as a subscript.

Also while here, add an attr_reader to SQLArray to read the array it
wraps.

In order to work correctly with the fact that you can specify column
aliases as hashes (a syntax I've never liked), modify the code to
transform the hashes to AliasedExpressions before adding them to
the opts, instead of handling them specially at SQL string creation.