Support savepoints when using SQLite
This commit makes the shared SQLite adapter use the savepoint
support used by the shared PostgreSQL and MySQL adapters.  It does
away with the BlockTransactions module, as only the SQLite and
Amalgalite adapters used that and it doesn't handle savepoints.
Since both the SQLite and Amalgalite adapters use the shared
SQLite adapter, they both support savepoints.