Make the tactical_eager_loading plugin not depend on the identity_map plugin
After giving it some thought, the tactical_eager_loading plugin
doesn't use the identity map at all, other than to not work when
there isn't an active identity map.  That seems like a pointless
restriction to me, so I've eliminated it.  The reason that it works
without the identity_map plugin is that it basically builds a type
of identity map internally which is used inside the :eager_loader
block.

The lazy_attributes plugin still depends on both, since it needs
the identity_map plugin and the tactical_eager_loading plugin in
order to load the lazy attribute for a bunch of models at once.