Fix graphing of datasets with dataset sources (Fixes #271)
This actually took a significant amount of internal restructing.

Dataset#from now does processing on the arguments to make
sure that the entires in the :from option are in a consistent
format.  This is similar to the change made recently to
Dataset#select.  This is mirrored by changes to Dataset#source_list
and Dataset#table_ref to not process the arguments.  This should
only break backwards compatibility if you were modifying the :from
option directly and using datasets or hashes.

Dataset#graph has been modified a little to handle using
an alias for datasets.  The implementation is a bit less optimal
than I would like, as it assumes specific behavior in Dataset#join,
but that's life.  It's possible that in the future Sequel will be
able to record all table aliases instead of just keeping track of
the number of dataset sources used.

Dataset#simple_select_all? is now only true if the only source in
the dataset is a Symbol, as previously datasets with a single
dataset source were considered simple select alls.

Dataset#first_source has been renamed to first_source_alias, with
an alias for the old name.  That's because it gives you the alias,
not the table/dataset used.

Dataset#dataset_alias private method was added.  It allows easier
overridding of the aliases used, incase the default aliases (tX)
are already used.

This commit removes Dataset#table_exists?, since it never worked
perfectly.  Database#table_exists? should be used instead.