Fix foreign key table constraints on MySQL
Sequel used to have a single method for handling column reference
SQL creation.  It was designed to be used for column constraints, and
the table constraints code also used it with some extra SQL.  The
MySQL adapter overrode it and made the method use the table
constraint format, since MySQL doesn't support the column constraint
format.  Unfortunately, when the table constraint code called it, the
it didn't generate the correct SQL.

This adds column_references_column_constraint_sql and
column_references_table_constraint_sql methods, both of which call
column_references_sql.  The MySQL adapter overrides
column_references_column_constraint_sql to use a table constraint.
This removes the aliasing of the default column_references_sql in
the MySQL adapter, since it is no longer needed.

This adds an integration test that for handling both types
of foreign_key calls in the create table block.  There is already
a test for add_foreign_key with a single key in an alter_table
block.  I also added a test for adding composite foreign keys in
an alter_table block.