Normalize bougyman's patches and improve the migrator specs
This commit makes bougyman's patches looks more like the rest of
Sequel's code.  Don't duplicate option defaults both in the default
argument and in the method body.  Instead, have a blank hash as the
default argument and use the RDoc to specify the defaults.

Raise a Sequel::Error instead of a string if an invalid directory
is given.

Pass option hash from the top level methods directly to the lower
level methods, instead of creating additional hashes.

Refactor and simplify some of the internal methods.

Change the specs to use a MockDatabase subclass instead of the
DummyMigrationDB.  Check the SQL produced to make sure that the
migrator is doing the correct thing.