Increase MySQL select performance using the native adapter
This optimizes the inner loop in the native MySQL adapter's
fetch_rows.  It eliminates some redundant calls and uses a
separate proc for tinyint so that the convert_type call is no
longer necessary.

I've benchmarked this and it is roughly an 18% performance
increase when selecting 1000 rows with 10 integer columns.