Significant increase in performance in the JDBC adapter
Using the same type of benchmark as in recent commits, this is
results in about a 10% increase in JDBC select performance.

JDBC performance could probably be way faster than it is
currently, but I'm not sure how to structure it like the
PostgreSQL and MySQL native adapters.  If you are a JRuby guru
and have knowledge of JDBC, I'd appreciate your help.