Majorly increase MySQL select performance using the native adapter
Instead of using a default proc that returns its argument, check
for the existence of a type proc and if there isn't one, just
assign the value.  Instead of an 18% performance increase as in the
previous commit, this yields around a 75% performance increase.

Also, the initial commit message in this series of performance
patches was wrong.  These recent benchmarks were done with selecting
1000 rows with 10 string columns, not 10 integer columns.  Integer
columns are probably significantly slower due to having to convert
the values.