Allow connection pool options to be specified in connection string
You can now specify the connection pool options directly in the
connection string:

   postgres:///database?max_connections=10&pool_timeout=20

An error will now be raised immediately if max_connections is not
positive.

This breaks backwards compatibility slightly as before you were
allowed to specify a Float for the pool_timeout.  If you do so
now, it gets explicitly converted to an Integer.