Allow single_threaded setting to be specified in connection string
You can now specify the single threaded section directly in the
connection string:

   postgres:///database?single_threaded=t
   postgres:///database?single_threaded=f

This breaks backwards compatibility slightly as before any value
other than nil or false would cause the single threaded setting
to be set.  Now, it uses the database's typecasting code to
determine whether the value should be considered true or false.

This is unlikely to have a significant effect, as Sequel defaults
to thread-safe behavior, and you only needed to specify the
single_threaded setting to turn the setting on.  However, I think
this new behavior is closer to expectations.