Remove UnsupportedIntersectExcept(All)? modules, replace with methods
This is the first in a series of commits that gets rid of the
adapters/utils directory.  It removes the modules that overrode the
methods in the dataset subclasses with two methods,
supports_intersect_except? and supports_intersect_except_all?,
and modifies the standard intersect and except methods to use
those methods to determine whether to raise an error.

This approach isn't as fast as the previous approach, but it allows
for more reflection support and is simpler to implement and use.