Remove SQLStandardDateFormat, replace with requires_sql_standard_datetimes? method
One change here is that before, ADO used SQL standard datetimes,
which was a mistake as ADO can be used to connect to any database.
It doesn't appear to be required on Microsoft SQL Server (the default
when using ADO), so I'm not sure why it was used in the first place.
If MSSQL does require it, it should be added to the shared adapter
and not the ADO adapter.