Fix thread-safety issue in stored procedure code
The stored procedure code wasn't thread safe before.  Trying
to run the same stored procedure in separate threads could
previously result in the arguments to the procedure getting
clobbered.

Now, when you call a stored procedure, it clones itself and
passes the arguments to the clone to run, which should be
thread-safe.

I was originally planning on moving the stored procedure
support out of adapters/utils, but considering it requires
specific adapter database support and is currently only
supported on MySQL and JDBC, I'm planning on leaving it in
adapter/utils for now.