Fix changing a column's name, type, or null status on MySQL when column has a string default
This large commit pulls in the database default to ruby default
convertor from the schema_dumper plugin and has schema parser
add a :ruby_default entry to all column schemas using the
default convertor.

This is necessary to make changing a column column's name, type,
or null status on MySQL when the column has a string default.  This
is because MySQL doesn't quote string defaults in DESCRIBE TABLE.

I don't really like this course of action, as the default conversion
code is very ugly, but there doesn't appear to be a better way to
fix things.

This does have the advantage of supplying defaults as ruby objects
for all parsed schemas, which I think could have many potential
uses.  For example, you could use it in a after_initialize method
in a model to automatically set defaults, so that Sequel::Model
operated more like ActiveRecord.