Add Dataset#ungraphed, to remove the splitting of results into subhashes or associated records
This is mostly useful when using eager_graph as a shortcut to setup
joins that would be tedious to create manually, but where you really
want to do some calculation or reporting and you don't want any
post-processing of the results.

It also works for plain datasets, where you can use it to have graph
automatically alias columns for you, but still want results in a
single hash instead of subhashes per table.