Add SQL::WindowFunction, expand virtual row blocks to support them and other constructions
This adds two new classes, SQL::WindowFunction and SQL::Window.
Window is separated from WindowFunction because some databases (at
least PostgreSQL) allow you to specify windows separately.

VirtualRow's method_missing now considers if a block was given to
the method.  If no block was used, the old behavior is used.  If
a block is given, the following new criteria are used:

* If there are no arguments, an SQL::Function with the name of
  method used, and no arguments.
* If the first argument is :*, an SQL::Function is created with a
  single wildcard argument (*).
* If the first argument is :distinct, an SQL::Function is created
  with the keyword DISTINCT prefacing all remaining arguments.
* If the first argument is :over, the second argument if provided
  should be a hash of options to pass to SQL::Window.  The options
  hash can also contain :*=>true to use a wildcard argument as the
  function argument, or :args=>... to specify an array of arguments
  to use as the function arguments.

So you can now use virtual row blocks for the following things:

  # Functions
  ds.select{version{}} # SELECT version()
  ds.select{count(:*){}} # SELECT count(*)
  ds.select{count(:distinct, col1){}} # SELECT count(DISTINCT col1)
  # Window Functions
  ds.select{rank(:over){}} # SELECT rank() OVER ()
  ds.select{count(:over, :*=>true){}} # SELECT count(*) OVER ()
  ds.select{sum(:over, :args=>col1, :partition=>col2, :order=>col3){}}
  # SELECT sum(col1) OVER (PARTITION BY col2 ORDER BY col3)

PostgreSQL can now use the Dataset#window method, which creates named
windows, which window functions can refer to using the :window
option.

Support for Window and WindowFunction was added to Dataset#qualify,
and support for PlaceholderLiteralString was also added.