Add support for Common Table Expressions, which use the SQL WITH clause
CTEs are supported by MSSQL 2005+, DB2 7+, Firebird 2.1+, Oracle 9+,
and PostgreSQL 8.4+.  They allow you to specify inline views that
the SELECT query can reference, and also support a recursive mode
that allows tables in the WITH clause to reference themselves. This
allows things not normally possible in standard SQL, like loading all
descendents in a tree structure in a single query.

The standard with clause takes an alias and a dataset:

  ds = DB[:vw]
  ds.with(:vw, DB[:table].filter{col < 1})
  # WITH vw AS (SELECT * FROM table WHERE col < 1) SELECT * FROM vw

The recursive with clause usage takes an alias, a nonrecursive
dataset, and a recursive dataset:

  ds.with_recursive(:vw,
    DB[:tree].filter(:id=>1),
    DB[:tree].join(:vw, :id=>:parent_id).
              select(:vw__id, :vw__parent_id))
  # WITH RECURSIVE vw AS (SELECT * FROM "tree"
  #     WHERE ("id" = 1)
  #     UNION ALL
  #     SELECT "vw"."id", "vw"."parent_id"
  #     FROM "tree"
  #     INNER JOIN "vw" ON ("vw"."id" = "tree"."parent_id"))
  # SELECT * FROM "vw"

I've only tested this on an PostgreSQL 8.4 release candidate, but
based on my research it should support MSSQL, DB2, and Firebird.
Oracle apparently uses a different method for recursive queries.

This commit contains some fixes for the window functions committed
previously.  Like now, it actually works on ruby 1.9.  The
Dataset#supports_window_functions? method is now defined and
defaults to false, and the adapters that support window
functions override it.  If you attempt to use a window function
on an unsupported dataset, an Error will be raised.

Sequel's default is to support common table expressions, so the
databases that don't support CTEs have been modified to not use the
with clause if they previously used the default clauses.

The MSSQL adapter was already using the :with option for setting
NOLOCK on the query, so I changed the setting to use :table_options.

Because the clauses the PostgreSQL adapter will use depends on the
server version, I changed the code for getting the server version
to use static SQL in order to avoid infinite recursion.

PostgreSQL dataset's explain and analyze methods were broken since
3.0, this commit fixes them and adds a simple spec.

In order to implement the with_recursive method easily, the compound
dataset methods (union, except, and intersect) now take an options
hash instead of a true/false flag.  The previous true/false flag is
still supported for backwards compatibility.  You can now specify it
with the :all option, and there is now a :from_self option that you
can set to false if you don't want to return a from_self dataset.

This commit adds the qualify and with dataset methods to
Sequel::Model.