More improvements for Microsoft SQL Server
This commit has a whole bunch of small improvments for MSSQL.  Now,
integration testing on MSSQL results in:

* ado: 117 examples, 23 failures
* jdbc: 119 examples, 14 failures
* odbc: 117 examples, 12 failures

For the ado adapter, Database#transaction now supports
Sequel::Rollback.  It also doesn't give an error if a dataset fetch
returns no rows.  The MSSQL subadapter now returns autoincremented
primary keys as integers instead of strings.

DateTime/Time formatting in the ODBC adapter was refactored and
should now be easier to understand.

The MSSQL shared adapter now uses the correct datetime and blob
formats.  It also uses the correct drop_index syntax.

Model#save now no longer includes primary key fields in the update
call.  This could potentially result in an error if a primary key
field is modified, but the previous behavior was probably incorrect.
Primary key fields should not be modified.

SQL::WindowFunction is now a subclass of GenericExpression instead
of Expression.  This allows it to be treated like any other function,
so you can alias it, cast it, etc..