Slight increase in ConnectionPool performance
This is a fairly simple refactoring that makes things a little more
verbose, but performs better.  I tried converting the ConnectionPool
to use a Monitor and ConditionVariable, but the performance was
actually worse than the original code (at least on ruby 1.8 and
jruby).

I benchmarked all three ways (original, ConditionVariable, and this
commit), and this commit was the fastest in every case (ruby 1.8,
ruby 1.9, and jruby).