Make the JDBC adapter accept a :convert_types option to turn off Java type conversion and double performance
This gives the user the choice to turn off conversion from Java types
to ruby types, in order to roughly double the performance when
fetching rows.  This can be set at both the database and dataset
level:

  DB = Sequel.jdbc('jdbc:postgresql://host/database',
                   :convert_types=>false)
  DB.convert_types = true
  ds = DB[:table]
  ds.convert_types = true

If you turn off Java type conversion, Date, DateTime, Time, and
BigDecimal fields will probably be returned as Java types, which
have a different API than ruby types.  I recommend only turning
it off if there is a bottleneck in your application, and then
only for the dataset that is the bottleneck.  Since Sequel allows
you to modify the setting at the dataset level, that's fairly easy
to do.

Note that the setting is not a database option like :from, its a
setting like quote_identifiers, and modifies the dataset directly
instead of returning a modified copy, so use with care.