Move convert_tinyint_to_bool method from Sequel to Sequel::MySQL
This is a backwards incompatible change, but I think it is for the
best.  Since the native MySQL adapter is the only adapter that
converts the tinyint type to boolean, it's the only adapter that
should have this method.  Other databases, such as MSSQL, have both
a tinyint type and another type for booleans (MSSQL uses bit), so
having the tinyint conversion on by default there is a bad idea.

In addition to moving the method, the implementation has changed a
bit as well.  Now, only tinyint(1) columns are considered boolean,
so if you have a tinyint(4) column (the default tinyint size), it
is not considered boolean.

Due to limitations in the native MySQL driver, all tinyint column
values are converted to booleans upon retrieval, not just tinyint(1)
column values.

Appropriately, Sequel now uses the tinyint(1) type for TrueClass and
FalseClass generic types.  Unfortunately, Sequel previously just used
tinyint instead of tinyint(1), so databases created with the previous
generic type support using TrueClass or FalseClass on MySQL databases
will now result in those columns being considered integer columns.

These changes could break applications.  Care should be taken when
upgrading.

This commit also adds support for the following types to the schema
parser and schema dumper: nchar, nvarchar, ntext, smalldatetime,
smallmoney, binary, and varbinary.