Support offsets on MSSQL using ROW_NUMBER window function
Since MSSQL doesn't support OFFSET natively, it has to be emulated
by running the query in a subquery and using the ROW_NUMBER window
function to only return the rows desired.  This is a bit ugly, but
does work (with integration tests added to check).

Note that MSSQL requires that you provide an order if you are going
to use the offset feature.  Most use cases for a limit with an offset
are going to provide an order, so I don't think this is too much of
an imposition.

Because there isn't a good way to not return the ROW_NUMBER field,
Dataset#each is modified to remove the ROW_NUMBER field if an
offset is used.

This commit adds Dataset#unlimited, similar to unfiltered and
unordered.

While here, comment the shared MSSQL adapter.