Model associations now default to associating to classes in the same scope (Fixes #274)
This changes the default for the :class_name option to use a model in
the same scope as the current model.  This is a slight breakage of
backwards compatibility, but should only affect users who are
associating a model inside a module to a model outside the module.
Defaulting to an object in the same scope leads to less surprising
behavior.

This patch also is a cleaner implementation, since the code no longer
varies per association type.