Add an assocation_proxies plugin that uses proxies for associations
My design philosophy with Sequel has been that proxies make the code
more difficult to understand, and should be avoided unless necessary.
I still feel that the fact that Sequel doesn't use proxies for
associations is a good thing.  However, Sequel does try to be
flexible, and this plugin shows how easy it is to modify Sequel to
use proxies for associations if you so desire.

The association proxies implemented by this commit are minimal in
terms of scope.  All they do is pass array methods to the array of
associated objects, and other methods to the association dataset.
This has the advantage that you can call dataset methods directly
on the association method, without all of the complexity entailed by
ActiveRecord's proxy implementation.  For example, you still need to
use the add/remove/remove_all methods on the object to modify the
association.  I'll consider patches to this plugin to add more
advanced behavior, but don't plan on adding additional support
myself.

Using this plugin should not break backwards compatibility except in
extreme cases.  One thing that could be broken by it is code that
expects NoMethodError when the object returned by the association
method is sent a method is the array of associated objects doesn't
respond to, but that the association's dataset does.  Another is code
that expects the association to be cached just by calling the
associated method, without calling any methods on the object
returned. Because the array methods are given priority over
the dataset methods, if you are just using array methods, this plugin
should not have an effect.

To ease implementation, the private class method
def_association_method has been split from
def_association_dataset_methods.  Also, the Sequel::BasicObject class
is added if ruby 1.9 is not being used.  Because of the way scoping
works in ruby, Sequel will use the ::Sequel::BasicObject class as the
base class for AssociationProxy and VirtualRow on 1.8, and the
::BasicObject class on 1.9.